Update option Menu
