+++ design_system/components/molecules/OptionButton/index.ts
@@ -1,0 +1,1 @@
+export { OptionButton } from './OptionButton';
